+++ soft/uemul/cpu.h
@@ -42,8 +42,8 @@ class CPU {
   void WriteRAM(uint16_t addr, uint8_t data);
   uint8_t ReadRAM(uint16_t addr);
 
-  void WriteVRAM(uint32_t addr, uint16_t data);
-  uint16_t ReadVRAM(uint32_t addr);
+  void WriteVRAM(uint16_t addr, uint16_t data);
+  uint16_t ReadVRAM(uint16_t addr);
 
   uint8_t ramp() { return ports[5]; }  // RAM page is stored in PORT5
